Implements creation and deletion of Items

+++ impl/src/main/java/dev/berlitz/demo/impl/repository/ItemRepository.java
@@ -3,6 +3,6 @@ package dev.berlitz.demo.impl.repository;
 import dev.berlitz.demo.impl.model.response.Item;
 import org.springframework.data.repository.CrudRepository;
 
-public interface ItemRepository extends CrudRepository<Item, Long> {
+public interface ItemRepository extends CrudRepository<Item, Integer> {
 
 }
